+++ src/main/java/projet/ais/repository/Niveau3PaysRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Niveau3Pays;
 
-public interface Niveau3PaysRepository  extends JpaRepository<Niveau3Pays, Long>{
+public interface Niveau3PaysRepository  extends JpaRepository<Niveau3Pays, Integer>{
     
 }
