+++ src/main/java/projet/ais/repository/ParametreGenerauxRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.ParametreGeneraux;
 
-public interface ParametreGenerauxRepository  extends JpaRepository<ParametreGeneraux, Long>{
+public interface ParametreGenerauxRepository  extends JpaRepository<ParametreGeneraux, Integer>{
     
 }
