+++ src/main/java/projet/ais/repository/RenvoieParametreRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.RenvoieParametre;
 
-public interface RenvoieParametreRepository extends JpaRepository<RenvoieParametre, Long>{
+public interface RenvoieParametreRepository extends JpaRepository<RenvoieParametre, Integer>{
     
 }
