+++ src/main/java/projet/ais/repository/UniteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Unite;
 
-public interface UniteRepository extends JpaRepository<Unite, Long>{
+public interface UniteRepository extends JpaRepository<Unite, Integer>{
     
 }
